Bug 1563467, ensure main navigation is keyboard accessible
