+++ kuma/javascript/src/header/header.jsx
@@ -1,13 +1,11 @@
 //@flow
 import * as React from 'react';
-import { useMemo } from 'react';
 import { css } from '@emotion/core';
 
-import { getLocale, gettext } from '../l10n.js';
+import { getLocale } from '../l10n.js';
 import Login from './login.jsx';
 import Logo from '../icons/logo.svg';
-import Dropdown from './dropdown.jsx';
-import { Row, Strut } from '../layout.jsx';
+import MainMenu from './main-menu.jsx';
 import Search from './search.jsx';
 
 import type { DocumentData } from '../document.jsx';
@@ -65,15 +63,6 @@ const styles = {
             display: 'block'
         }
     }),
-    menus: css({
-        gridArea: 'M',
-        flexWrap: 'wrap',
-        fontSize: 15,
-        fontWeight: 'bold',
-        button: {
-            lineHeight: '32px'
-        }
-    }),
     search: css({
         gridArea: 'S',
         justifySelf: 'stretch'
@@ -92,118 +81,6 @@ type Props = {
 export default function Header(props: Props): React.Node {
     const locale = getLocale();
 
-    // The menus array includes objects that define the set of
-    // menus displayed by this header component. The data structure
-    // is defined here during rendering so that the gettext() calls
-    // happen after the string catalog is available. And we're using
-    // useMemo() so that localization only happens when the locale
-    // changes.
-    const menus = useMemo(
-        () => [
-            {
-                label: gettext('Technologies'),
-                url: `/${locale}/docs/Web`,
-                items: [
-                    { url: `/${locale}/docs/Web/HTML`, label: gettext('HTML') },
-                    { url: `/${locale}/docs/Web/CSS`, label: gettext('CSS') },
-                    {
-                        url: `/${locale}/docs/Web/JavaScript`,
-                        label: gettext('JavaScript')
-                    },
-                    {
-                        url: `/${locale}/docs/Web/Guide/Graphics`,
-                        label: gettext('Graphics')
-                    },
-                    { url: `/${locale}/docs/Web/HTTP`, label: gettext('HTTP') },
-                    {
-                        url: `/${locale}/docs/Web/API`,
-                        label: gettext('APIs / DOM')
-                    },
-                    {
-                        url: `/${locale}/docs/Mozilla/Add-ons/WebExtensions`,
-                        label: gettext('Browser Extensions')
-                    },
-                    {
-                        url: `/${locale}/docs/Web/MathML`,
-                        label: gettext('MathML')
-                    }
-                ]
-            },
-            {
-                label: gettext('References & Guides'),
-                url: `/${locale}/docs/Learn`,
-                items: [
-                    {
-                        url: `/${locale}/docs/Learn`,
-                        label: gettext('Learn web development')
-                    },
-                    {
-                        url: `/${locale}/docs/Web/Tutorials`,
-                        label: gettext('Tutorials')
-                    },
-                    {
-                        url: `/${locale}/docs/Web/Reference`,
-                        label: gettext('References')
-                    },
-                    {
-                        url: `/${locale}/docs/Web/Guide`,
-                        label: gettext('Developer Guides')
-                    },
-                    {
-                        url: `/${locale}/docs/Web/Accessibility`,
-                        label: gettext('Accessibility')
-                    },
-                    {
-                        url: `/${locale}/docs/Games`,
-                        label: gettext('Game development')
-                    },
-                    {
-                        url: `/${locale}/docs/Web`,
-                        label: gettext('...more docs')
-                    }
-                ]
-            },
-            {
-                label: gettext('Feedback'),
-                url: `/${locale}/docs/MDN/Feedback`,
-                items: [
-                    {
-                        url: 'https://support.mozilla.org/',
-                        label: gettext('Get Firefox help'),
-                        external: true
-                    },
-                    {
-                        url: 'https://stackoverflow.com/',
-                        label: gettext('Get web development help'),
-                        external: true
-                    },
-                    {
-                        url: `/${locale}/docs/MDN/Community`,
-                        label: gettext('Join the MDN community')
-                    },
-                    {
-                        label: gettext('Report a content problem'),
-                        external: true,
-                        url:
-                            'https://github.com/mdn/sprints/issues/new?template=issue-template.md&projects=mdn/sprints/2&labels=user-report&title={{PATH}}'
-                    },
-                    {
-                        label: gettext('Report a bug'),
-                        external: true,
-                        url: 'https://bugzilla.mozilla.org/form.mdn'
-                    }
-                ]
-            }
-        ],
-        [locale]
-    );
-
-    // One of the menu items has a URL that we need to substitute
-    // the current document path into. Compute that now.
-    let path = encodeURIComponent(
-        `/${locale}` + (props.document ? `/docs/${props.document.slug}` : '')
-    );
-
     return (
         <div css={styles.header}>
             <a
@@ -219,33 +96,7 @@ export default function Header(props: Props): React.Node {
                 // stylesheets define a bunch of styles on <hr>.
             }
             <div css={styles.rule} />
-            <Row css={styles.menus}>
-                {menus.map((m, index) => (
-                    <React.Fragment key={index}>
-                        <Dropdown label={<a href={m.url}>{m.label}</a>}>
-                            {m.items.map((item, index) => (
-                                <li key={index}>
-                                    {item.external ? (
-                                        <a
-                                            target="_blank"
-                                            rel="noopener noreferrer"
-                                            href={item.url.replace(
-                                                '{{PATH}}',
-                                                path
-                                            )}
-                                        >
-                                            {item.label} &#x1f310;
-                                        </a>
-                                    ) : (
-                                        <a href={item.url}>{item.label}</a>
-                                    )}
-                                </li>
-                            ))}
-                        </Dropdown>
-                        {index < menus.length - 1 && <Strut width={16} />}
-                    </React.Fragment>
-                ))}
-            </Row>
+            <MainMenu document={props.document} />
             <div css={styles.search}>
                 <Search initialQuery={props.searchQuery || ''} />
             </div>
