+++ kuma/static/styles/components/structure/page-header/_vars_nav.scss
@@ -1,0 +1,5 @@
+$main-nav-z: 95;
+$main-nav-box-shadow: 0 2px 8px 0 rgba(12, 12, 13, .1);
+$main-menu-border: solid 1px #d8dfe2;
+$main-menu-border-radius: 4px;
+$main-menu-item-background-color: #3d7e9a;
