+++ kuma/static/styles/components/structure/page-header/nav-main.scss
@@ -1,4 +1,6 @@
 @import '_mixin_nav';
+@import '../../../includes/_vars';
+@import '_vars_nav';
 
 /*
 Main navigation menu
@@ -44,5 +46,103 @@ html:not(.no-js) .nav-main-item > a {
 
     @media #{$mq-small-desktop-and-up} {
         @include large-inline-nav-item();
+    }
+}
+
+/*
+New main navigation menu
+********************************************************************** */
+.main-nav {
+    grid-area: M;
+    @include set-font-size(18px);
+
+    & > ul {
+        display: block;
+
+        @media #{$mq-tablet-and-up} {
+            display: flex;
+            position: relative;
+            justify-content: space-evenly;
+            font-weight: bold;
+            z-index: $main-nav-z;
+        }
+
+        .main-menu-arrow {
+            padding-left: 2px;
+            font-size: 75%;
+        }
+    }
+
+    li ul {
+        display: none;
+        background-color: #fff;
+        margin: 0;
+        padding: 4px 0;
+
+        @media #{$mq-tablet-and-up} {
+            box-shadow: $main-nav-box-shadow;
+            border: $main-menu-border;
+            border-radius: $main-menu-border-radius;
+        }
+
+        a {
+            display: block;
+            padding: 6px 16px;
+            width: 100%;
+
+            &:hover {
+                background-color: $main-menu-item-background-color;
+                color: #fff;
+                text-decoration: none;
+            }
+
+            @media #{$mq-tablet-and-up} {
+                width: unset;
+            }
+        }
+
+        li {
+            width: 100%;
+
+            @media #{$mq-tablet-and-up} {
+                width: unset;
+            }
+        }
+    }
+
+    .top-level-entry-container {
+        position: relative;
+
+        ul[aria-hidden='true'] {
+            display: none;
+        }
+
+        ul[aria-hidden='false'] {
+            display: block;
+        }
+
+        @media #{$mq-tablet-and-up} {
+            &:hover ul,
+            &:focus-within ul {
+                display: inherit;
+
+                @media #{$mq-tablet-and-up} {
+                    display: block;
+                    position: absolute;
+                    width: max-content;
+                }
+            }
+        }
+
+        button {
+            border: none;
+            width: 100%;
+            text-align: left;
+            padding: 10px;
+
+            @media #{$mq-tablet-and-up} {
+                width: unset;
+            }
+        }
     }
 }
