+++ kuma/static/styles/includes/_vars.scss
@@ -307,7 +307,6 @@ $submenu-z: 99;
 $logo-z: 97; /* needs to be above main-nav */
 $toolbox-z: 97; /* needs to be above main-nav */
 $main-header-z: 96;
-$main-nav-z: 95;
 $document-actions-z: 89;
 $document-head-z: 88;
 $toc-z: 79;
